Added Javadoc.

+++ app/src/main/java/edu/cnm/deepdive/nmmedicalcannabis/entities/CardDatabase.java
@@ -6,6 +6,9 @@ import com.j256.ormlite.table.DatabaseTable;
 import java.sql.Timestamp;
 import java.util.Date;
 
+/**
+ * Creates database for the patient card information
+ */
 @DatabaseTable(tableName = "PATIENT_CARD")
 public class CardDatabase {
 
@@ -26,53 +29,102 @@ public class CardDatabase {
   private  Date expDate;
 
   @DatabaseField(columnName = "UNITS_AVAILABLE", canBeNull = false)
-  private int unitsAvailable;
+  private double unitsAvailable;
 
+  /**
+   * Creates the key column and issues an ID starting at 1.
+   * @return ID Key
+   */
   public int getId() {
     return id;
   }
 
+  /**
+   * Creates access to ID Key
+   * @param id KEY ID
+   */
   public void setId(int id) {
     this.id = id;
   }
 
+  /**
+   * Creates a timestamp when each transaction is made.
+   * @return returns timestamp
+   */
   public Timestamp getCreated() {
     return created;
   }
 
+  /**
+   * Creates access to timestamp
+   * @param created the date and time transaction was created
+   */
   public void setCreated(Timestamp created) {
     this.created = created;
   }
 
+  /**
+   * Gets the ID number of the patient card
+   * @return returns card number
+   */
   public long getCardID() {
     return cardID;
   }
 
+  /**
+   * Gives access to the card ID number
+   * @param cardID the number of the patient card
+   */
   public void setCardID(long cardID) {
     this.cardID = cardID;
   }
 
+  /**
+   * Gets the date of when the card was issued.
+   * @return returns the issue date
+   */
   public Date getIssueDate() {
     return issueDate;
   }
 
+  /**
+   * Gives access to the issue date
+   * @param issueDate
+   */
   public void setIssueDate(Date issueDate) {
     this.issueDate = issueDate;
   }
 
+  /**
+   * Gets the expiration date of the card. One year from the issue date.
+   * @return returns expiration date.
+   */
   public Date getExpDate() {
     return expDate;
   }
 
+  /**
+   * Gives access to the expiration date
+   * @param expDate
+   */
   public void setExpDate(Date expDate) {
     this.expDate = expDate;
   }
 
-  public int getUnitsAvailable() {
+  /**
+   * Gets the units available from when the users starts using the app
+   * or if the user is a new patient
+   * @return returns the units available at the time of download.
+   */
+  public double getUnitsAvailable() {
     return unitsAvailable;
   }
 
-  public void setUnitsAvailable(int unitsAvailable) {
+  /**
+   * Gives access to the units available.
+   * @param unitsAvailable
+   */
+  public void setUnitsAvailable(double unitsAvailable) {
     this.unitsAvailable = unitsAvailable;
   }
